Change mode-matching algorithm
This new method allows for different numbers of modes between waveguides
and is formulated in a way to always use the scalar product from small
to bigger waveguide. This has the huge advantage, that additional
reflections caused by the H-field not being 0 at a boundary that is
being reduced don't have to be taken into account.